@LoserCheems 2025-08-12
## Description
Fix numerical equivalence in the split-KV forward path of Dynamic Mask Attention #47 . The root cause was double-scaling the softmax LSE (multiplying row_max by softmax_scale again during normalization), which led to underestimated merge weights in the split-combine kernel and produced outputs with compressed magnitude. After the fix, CUDA matches the Python reference precisely.

## Type of Change
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] CUDA kernel improvement
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation update
- [ ] Performance optimization
- [ ] Code refactoring

## Related Issues
- Fixes: N/A
- Related to: N/A

## Changes Made
Remove the redundant softmax_scale factor from LSE computation so that per-split LSE is consistent with logits scaling and the split-KV combine kernel can reweight outputs correctly.

### Code Changes
- [x] Updated CUDA kernels
  - softmax.h
    - Change:
      ```
      lse(mi) = row_max(mi) * softmax_scale + __logf(sum);
      ```
      to:
      ```
      lse(mi) = row_max(mi) + __logf(sum);
      ```
- [ ] Modified Python API
- [ ] Changed build system
- [ ] Updated dependencies

### Documentation
- [ ] Updated README
- [ ] Updated API documentation
- [ ] Added examples
- [ ] Updated benchmarks

## Testing
Validated numerics and performance locally and in container.

- [x] Existing benchmarks pass: `python benchmarks/forward_equivalence.py --test-type cuda`
- [x] Benchmarks show no performance regression
- [ ] Added new tests
- [ ] Tested on multiple GPU architectures (SM 8.9 validated)

### Test Configuration
- OS: Ubuntu (Docker)
- Python: 3.10+
- PyTorch: 2.8.0a0+5228986c39.nv25.05
- CUDA: compatible with RTX 4090 driver
- GPU: NVIDIA GeForce RTX 4090 (SM 8.9)

## Performance Impact
Pure numerical fix. No extra memory traffic or compute. Performance unchanged or slightly improved by removing redundant math.

### Before
- Config: b=1, h=2, kv_h=1, q_len=256, k_len=512, d=128, causal=True
  - Python: 299.66 ms
  - CUDA: 3.04 ms
  - Numerics: Fail (compressed range)

### After
- Config: b=1, h=2, kv_h=1, q_len=256, k_len=512, d=128, causal=True
  - Python: 274.12 ms
  - CUDA: 1.94 ms (≈141x)
  - Numerics: Pass (all elements within tolerance)
- Config: b=1, h=2, kv_h=1, q_len=64, k_len=512, d=128, causal=True
  - Python: 268.22 ms
  - CUDA: 1.46 ms (≈184x)
  - Numerics: Pass

## Breaking Changes
None. Public API and expected outputs remain unchanged; only LSE computation is corrected.

## Checklist
- [x] My code follows the project's style guidelines
- [x] I have performed a self-review
- [x] I have commented hard-to-understand areas
- [x] My changes introduce no new warnings
- [x] New and existing benchmarks pass locally

### CUDA-specific (if applicable)
- [x] CUDA kernels compile without warnings
- [x] Tested on SM 8.0+ (validated on SM 8.9)
- [x] Memory usage unchanged; no leaks

## Additional Notes
- Non-split path normalizes O within a single kernel and does not depend on stored LSE; split-KV must use per-split LSE (or m_i, l_i) to reweight and combine partial outputs stably. This fix aligns LSE with the softmax normalization constant.- [ ] Modified Python API
- [ ] Changed build system
- [ ] Updated dependencies

### Documentation
- [ ] Updated README
- [ ] Updated API documentation
- [ ] Added examples
- [ ] Updated benchmarks

## Testing
Validated numerics and performance locally and in container.

- [x] Existing benchmarks pass: `python benchmarks/forward_equivalence.py --test-type cuda`
- [x] Benchmarks show no performance regression
- [ ] Added new tests
- [ ] Tested on multiple GPU architectures (SM 8.9 validated)

### Test Configuration
- OS: Ubuntu (Docker)
- Python: 3.10+
- PyTorch: 2.8.0a0+5228986c39.nv25.05
- CUDA: compatible with RTX 4090 driver
- GPU: NVIDIA GeForce RTX 4090 (SM 8.9)

## Performance Impact
Pure numerical fix. No extra memory traffic or compute. Performance unchanged or slightly improved by removing redundant math.

### Before
- Config: b=1, h=2, kv_h=1, q_len=256, k_len=512, d=128, causal=True
  - Python: 299.66 ms
  - CUDA: 3.04 ms
  - Numerics: Fail (compressed range)

### After
- Config: b=1, h=2, kv_h=1, q_len=256, k_len=512, d=128, causal=True
  - Python: 274.12 ms
  - CUDA: 1.94 ms (≈141x)
  - Numerics: Pass (all elements within tolerance)
- Config: b=1, h=2, kv_h=1, q_len=64, k_len=512, d=128, causal=True
  - Python: 268.22 ms
  - CUDA: 1.46 ms (≈184x)
  - Numerics: Pass

## Breaking Changes
None. Public API and expected outputs remain unchanged; only LSE computation is corrected.

## Checklist
- [x] My code follows the project's style guidelines
- [x] I have performed a self-review
- [x] I have commented hard-to-understand areas
- [x] My changes introduce no new warnings
- [x] New and existing benchmarks pass locally

### CUDA-specific (if applicable)
- [x] CUDA kernels compile without warnings
- [x] Tested on SM 8.0+ (validated on SM 8.9)
- [x] Memory usage unchanged; no leaks

## Additional Notes
- Non-split path normalizes O within a single kernel and does not depend on stored LSE; split-KV must use per-split LSE (or m_i, l_i) to reweight and combine partial outputs stably. This fix aligns LSE with the softmax normalization constant.